sync_proxy_ldap/pyproject.toml
[project]
name = "sync_proxy_ldap"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "ldap3>=2.9.1",
    "python-dotenv>=1.1.1",
]
